move the LRU cache tests from decorators_test to core_test
